feat(backend): change the type of seen progress column

+++ libs/database/src/migrations/m20230419_create_seen.rs
@@ -42,7 +42,7 @@ impl MigrationTrait for Migration {
                     )
                     .col(
                         ColumnDef::new(Seen::Progress)
-                            .integer()
+                            .decimal()
                             .not_null()
                             .default(0),
                     )
